# docs/swagger.yaml
basePath: /
definitions:
...
  models.BookmarkRequest:
    properties:
      career_id:
        type: integer
    type: object
  models.CreateChallengeRequest:
    properties:
      career_id:
        type: integer
    type: object
  models.CreateProjectIdeaRequest:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Career discovery API for Nepal
  title: Carevo API
  version: "1.0"
paths:
...
  /careers/{id}/projects:
    get:
      parameters:
      - description: Career ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Browse project ideas for a career
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Career ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project idea
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateProjectIdeaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit a project idea
      tags:
      - projects
...
  /challenges:
    post:
      consumes:
      - application/json
      parameters:
      - description: Career ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateChallengeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start a 30-day challenge
      tags:
      - challenges
  /challenges/{id}/checkin:
    post:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Daily check-in for a challenge
      tags:
      - challenges
  /challenges/{id}/progress:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get challenge progress
      tags:
      - challenges
  /challenges/active:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get active challenge
      tags:
      - challenges
...
  /users/me/bookmarks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List bookmarked careers
      tags:
      - bookmarks
    post:
      consumes:
      - application/json
      parameters:
      - description: Career ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.BookmarkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Bookmark a career
      tags:
      - bookmarks
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"